# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workshop Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs
      parameters:
        - { name: workerId, in: query, schema: { type: integer } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Job" }
    post:
      operationId: createJob
      tags: [jobs]
      summary: Record a job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateJobBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
  /jobs/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    delete:
      operationId: deleteJob
      tags: [jobs]
      responses:
        "204": { description: Deleted }
  /jobs/{id}/expenses:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      operationId: updateJobExpenses
      tags: [jobs]
      summary: Replace all expense lines for a job and recalculate payouts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expenseLines]
              properties:
                expenseLines:
                  type: array
                  items:
                    type: object
                    required: [description, amount]
                    properties:
                      description: { type: string }
                      amount: { type: number }
                      paidByWorkerId: { type: integer, nullable: true }
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
...
components:
  schemas:
...
    JobExpenseLine:
      type: object
      required: [description, amount]
      properties:
        id: { type: integer }
        description: { type: string }
        amount: { type: number }
        paidByWorkerId: { type: integer, nullable: true, description: "If set, this worker paid the expense out of pocket and will be reimbursed on top of their split" }
        paidByWorkerName: { type: string, nullable: true }
    JobWorkerShare:
      type: object
      required: [id, jobId, workerId, workerName, amount]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        workerId: { type: integer }
        workerName: { type: string }
        amount: { type: number }
    Job:
      type: object
      required: [id, jobType, source, paymentMethod, grossAmount, cardFeeAmount, expensesTotal, netAmount, workerShare, workshopShare, vatPaidByCustomer, occurredAt]
      properties:
        id: { type: integer }
        jobType: { type: string, enum: [single, shared], description: "single = one worker, shared = multiple workers" }
        workerId: { type: integer, nullable: true }
        workerName: { type: string, nullable: true }
        cashReceivedByWorkerId: { type: integer, nullable: true, description: "Worker who physically received the cash payment from the customer" }
        cashReceivedByName: { type: string, nullable: true, description: "Name of the worker who received the cash" }
        source: { type: string, description: "Source of income / customer / job description" }
        paymentMethod: { type: string, enum: [cash, card] }
        grossAmount: { type: number, description: "Total amount paid by customer" }
        cardFeeAmount: { type: number, description: "VAT amount removed (only for card when VAT not paid by customer)" }
        vatPaidByCustomer: { type: boolean, description: "If true (card only), customer paid VAT separately so no VAT is removed from gross" }
        expensesTotal: { type: number, description: "Sum of job expense lines (parts, commission, etc)" }
        netAmount: { type: number, description: "Splittable amount = gross - VAT - expenses" }
        workerShare: { type: number }
        workshopShare: { type: number }
        notes: { type: string, nullable: true }
        expenseLines:
          type: array
          items: { $ref: "#/components/schemas/JobExpenseLine" }
        workerShares:
          type: array
          items: { $ref: "#/components/schemas/JobWorkerShare" }
        occurredAt: { type: string, format: date-time }
    CreateJobBody:
      type: object
      required: [source, paymentMethod, grossAmount]
      properties:
        jobType: { type: string, enum: [single, shared], description: "single = one worker, shared = multiple workers split manually" }
        workerId: { type: integer, description: "Required when jobType = single" }
        cashReceivedByWorkerId: { type: integer, nullable: true, description: "Cash jobs only: which worker physically received the cash from the customer" }
        workerShares:
          type: array
          description: "Required when jobType = shared. Each entry specifies a worker and their manual amount."
          items:
            type: object
            required: [workerId, amount]
            properties:
              workerId: { type: integer }
              amount: { type: number }
        source: { type: string }
        paymentMethod: { type: string, enum: [cash, card] }
        grossAmount: { type: number, description: "Total amount paid by customer" }
        vatPaidByCustomer: { type: boolean, description: "Card only. true = customer paid VAT on top (workshop keeps full gross). false = VAT included in gross and is removed before splitting." }
        expenseLines:
          type: array
          items:
            type: object
            required: [description, amount]
            properties:
              description: { type: string }
              amount: { type: number }
              paidByWorkerId: { type: integer, nullable: true, description: "Worker who paid this expense out of pocket" }
        notes: { type: string }
        occurredAt: { type: string, format: date-time }
...